fix a few broken links
Looks like the link is broken in
https://monix.io/docs/current/eval/task.html#comparison-with-the-scalaz-task. Even
though the link is properly rendered, it links to
`docs/current/best-practices/blocking.md` instead of
`docs/current/best-practices/blocking.html`.

Not having the new line in the link seems to fix it when tested locally.

Also fix one link in the readme.